Port AMP request signer to AWS SDK2
