--- flink-connector-prometheus-request-signer-amp/src/main/java/org/apache/flink/connector/prometheus/sink/aws/AWS4SignerForAuthorizationHeader.java
@@ -1,124 +1,0 @@
-/*
- *  Licensed to the Apache Software Foundation (ASF) under one or more
- *  contributor license agreements.  See the NOTICE file distributed with
- *  this work for additional information regarding copyright ownership.
- *  The ASF licenses this file to You under the Apache License, Version 2.0
- *  (the "License"); you may not use this file except in compliance with
- *  the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.apache.flink.connector.prometheus.sink.aws;
-
-import com.amazonaws.util.BinaryUtils;
-
-import java.net.URL;
-import java.util.Date;
-import java.util.Map;
-
-/**
- * AWS4 signer generating the 'Authorization' header based on the request body, parameters and AWS
- * credentials.
- */
-public class AWS4SignerForAuthorizationHeader extends AWS4SignerBase {
-
-    public AWS4SignerForAuthorizationHeader(
-            URL endpointUrl, String httpMethod, String serviceName, String regionName) {
-        super(endpointUrl, httpMethod, serviceName, regionName);
-    }
-
-    /**
-     * Computes an AWS4 signature for a request, ready for inclusion as an 'Authorization' header.
-     *
-     * @param headers The request headers; 'Host' and 'X-Amz-Date' will be added to this set.
-     * @param queryParameters Any query parameters that will be added to the endpoint. The
-     *     parameters should be specified in canonical format.
-     * @param bodyHash Precomputed SHA256 hash of the request body content; this value should also
-     *     be set as the header 'X-Amz-Content-SHA256' for non-streaming uploads.
-     * @param awsAccessKey The user's AWS Access Key.
-     * @param awsSecretKey The user's AWS Secret Key.
-     * @return The computed authorization string for the request. This value needs to be set as the
-     *     header 'Authorization' on the subsequent HTTP request.
-     */
-    public String computeSignature(
-            Map<String, String> headers,
-            Map<String, String> queryParameters,
-            String bodyHash,
-            String awsAccessKey,
-            String awsSecretKey,
-            String sessionToken) {
-        // first get the date and time for the subsequent request, and convert
-        // to ISO 8601 format for use in signature generation
-        Date now = new Date();
-        String dateTimeStamp = dateTimeFormat.format(now);
-
-        // update the headers with required 'x-amz-date' and 'host' values
-        headers.put("x-amz-date", dateTimeStamp);
-        if (sessionToken != null && !sessionToken.isEmpty()) {
-            headers.put("x-amz-security-token", sessionToken);
-        }
-
-        String hostHeader = endpointUrl.getHost();
-        int port = endpointUrl.getPort();
-        if (port > -1) {
-            hostHeader.concat(":" + port);
-        }
-        headers.put("Host", hostHeader);
-
-        // canonicalize the headers; we need the set of header names as well as the
-        // names and values to go into the signature process
-        String canonicalizedHeaderNames = getCanonicalizeHeaderNames(headers);
-        String canonicalizedHeaders = getCanonicalizedHeaderString(headers);
-
-        // if any query string parameters have been supplied, canonicalize them
-        String canonicalizedQueryParameters = getCanonicalizedQueryString(queryParameters);
-
-        // canonicalize the various components of the request
-        String canonicalRequest =
-                getCanonicalRequest(
-                        endpointUrl,
-                        httpMethod,
-                        canonicalizedQueryParameters,
-                        canonicalizedHeaderNames,
-                        canonicalizedHeaders,
-                        bodyHash);
-
-        // construct the string to be signed
-        String dateStamp = dateStampFormat.format(now);
-        String scope = dateStamp + "/" + regionName + "/" + serviceName + "/" + TERMINATOR;
-        String stringToSign =
-                getStringToSign(SCHEME, ALGORITHM, dateTimeStamp, scope, canonicalRequest);
-
-        // compute the signing key
-        byte[] kSecret = (SCHEME + awsSecretKey).getBytes();
-        byte[] kDate = sign(dateStamp, kSecret, "HmacSHA256");
-        byte[] kRegion = sign(regionName, kDate, "HmacSHA256");
-        byte[] kService = sign(serviceName, kRegion, "HmacSHA256");
-        byte[] kSigning = sign(TERMINATOR, kService, "HmacSHA256");
-        byte[] signature = sign(stringToSign, kSigning, "HmacSHA256");
-
-        String credentialsAuthorizationHeader = "Credential=" + awsAccessKey + "/" + scope;
-        String signedHeadersAuthorizationHeader = "SignedHeaders=" + canonicalizedHeaderNames;
-        String signatureAuthorizationHeader = "Signature=" + BinaryUtils.toHex(signature);
-
-        String authorizationHeader =
-                SCHEME
-                        + "-"
-                        + ALGORITHM
-                        + " "
-                        + credentialsAuthorizationHeader
-                        + ", "
-                        + signedHeadersAuthorizationHeader
-                        + ", "
-                        + signatureAuthorizationHeader;
-
-        return authorizationHeader;
-    }
-}
